Simplify component render method
Component's render was very complicated, and didn't work for users
who wanted to take control of rendering logic where they first render
components manually and then insert elements manually into the DOM.
This was due to the wrong assumption that "if component has an element,
then it must be in the DOM". In this case users wouldn't see their
components being rendered into the DOM despite claiming to be rendered.

This change fixes this by simplifying rendering logic to insert a
component's element to the root element on the render call. However, this
can result in multiple, unnecessary DOM insertions when the child is
already in the place where a user wants it to be. Unfortunately, browsers
don't optimize for el.insertBefore(child) where child is already the last
element of the el.

Also, previously this didn't take into account manual renders to a
different place within parent. This implementation also fixes that in a
performant way.