Fix component event declarations in parent components
TabView had an [issue](https://github.com/dashersw/erste/issues/36) where
it won't switch tabs. This was mainly due to the fact that we expected
events property to live on the final prototype. However, it could very
well live on parent prototypes. This caused an issue with binding the
wrong (parent's) this value to the handler.